requirements.txt
fastapi
uvicorn
python-docx
PyPDF2
transformers
torch # Or tensorflow
accelerate
scikit-learn
numpy